requirements.txt
Flask
Flask-SocketIO
stripe
Flask-SQLAlchemy
psycopg[binary]
Werkzeug
python-dotenv
openai
gunicorn
requests
Flask-Migrate
Flask-WTF
Pillow
ImageHash
cryptography
redis
google-auth
google-cloud-core
